portal/requirements.txt
bootstrap-flask
click
click-repl
cryptography
dominate
flake8
flask
flask-cors
flask-login
flask-migrate
flask-moment
flask_nav
flask_sessionstore
flask-sqlalchemy
flask-wtf
gunicorn
itsdangerous
jinja2
prompt_toolkit
pyjwt
pypng
pyyaml
markupsafe
# markupsafe==2.0.0rc1
onetimepass
psycopg2
psycopg2-binary
pyqrcode
redis
requests
six
sqlalchemy
# systemd-python  # requires pkg-config libsystemd-dev
systemd  # requires pkg-config libsystemd-dev
visitor
werkzeug
wheel
wtforms
